mt_workers/requirements.txt
amqp==2.2.1
Babel==2.5.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
chardet==3.0.4
cloudinary==1.8.0
Django==1.11.5
django-health-check==2.4.0
django-phonenumber-field==1.3.0
django-twilio==0.8.0
elasticsearch==5.4.0
flower==0.9.2
idna==2.6
kombu==4.1.0
mock==2.0.0
pbr==3.1.1
phonenumberslite==8.8.1
psycopg2==2.7.3.1
PyJWT==1.5.3
PySocks==1.6.7
python-http-client==3.0.0
pytz==2017.2
requests==2.18.4
sendgrid==3.6.5
sendgrid-django==4.0.4
six==1.11.0
tornado==4.5.2
twilio==6.6.3
urllib3==1.22
vine==1.1.4
websocket-client==0.44.0
